j'ai rendu mon code plus moche

+++ cogs/BombParty.py
@@ -17,10 +17,16 @@ class BombParty(commands.Cog):
     @commands.Cog.listener()
     async def on_guild_join(self, guild):
         await guild.create_role(name="Bomb Party Admin")
-        await guild.create_role(name="BP Current Player")
         # await guild.invoke(createChannel) #FIX THIS PLS
         
     ################ SETTINGS RELATED ################
+    @staticmethod
+    def isAdmin(Member):
+        for role in Member.roles:
+            if role.name == "Bomb Party Admin":
+                return True
+        return False
+    
     @staticmethod
     def getSettings(GuildId):
         file_name = './settings/' + str(GuildId) + '.json'
@@ -40,29 +46,32 @@ class BombParty(commands.Cog):
 
     @commands.command()
     async def showSettings(self, ctx: commands.Context):
-        settings = self.getSettings(ctx.guild.id)
-        await ctx.send(f'Language: {settings["Language"]}\nMinimum timing of the bomb: {settings["MinimumTiming"]} seconds')
+        if self.isAdmin(ctx.author):
+            settings = self.getSettings(ctx.guild.id)
+            await ctx.send(f'Language: {settings["Language"]}\nMinimum timing of the bomb: {settings["MinimumTiming"]} seconds')
 
     @commands.command()
     async def setLanguage(self, ctx: commands.Context, arg):
-        settings = self.getSettings(ctx.guild.id)
-        if arg in self._languages:
-            settings["Language"] = arg
-            setSettings(ctx.guild.id, settings)
-            await ctx.send(f'Language have been set to {arg}')
-        else:
-            await ctx.send("This language isn't accepted, or maybe you misstyped it...")
+        if self.isAdmin(ctx.author):
+            settings = self.getSettings(ctx.guild.id)
+            if arg in self._languages:
+                settings["Language"] = arg
+                setSettings(ctx.guild.id, settings)
+                await ctx.send(f'Language have been set to {arg}')
+            else:
+                await ctx.send("This language isn't accepted, or maybe you misstyped it...")
 
     @commands.command()
     async def setTiming(self, ctx: commands.Context, arg):
-        settings = self.getSettings(ctx.guild.id)
-        arg = int(arg)
-        if arg <= 10 and arg >= 0:
-            settings["MinimumTiming"] = arg
-            setSettings(ctx.guild.id, settings)
-            await ctx.send(f'Time have been set to {arg} seconds')
-        else:
-            await ctx.send("You can only chose a number between 0 and 10.")
+        if self.isAdmin(ctx.author):
+            settings = self.getSettings(ctx.guild.id)
+            arg = int(arg)
+            if arg <= 10 and arg >= 0:
+                settings["MinimumTiming"] = arg
+                setSettings(ctx.guild.id, settings)
+                await ctx.send(f'Time have been set to {arg} seconds')
+            else:
+                await ctx.send("You can only chose a number between 0 and 10.")
     ############# ENDOF SETTINGS RELATED #############
 
     ################ CHANNELS RELATED ################
@@ -86,39 +95,41 @@ class BombParty(commands.Cog):
 
     @commands.command()
     async def createChannel(self, ctx: commands.Context, arg=1):
-        arg = int(arg)
-        if arg <= 50 and arg >=1:
-            channels = getChannelNamesList(ctx.guild)
-            i = getMaxChannel(channels)
-            for _ in range(arg):
-                await ctx.guild.create_text_channel(self._channelPrefix + str(i))
-                i += 1
-            await ctx.send(f'{arg} channel(s) have been created.')
-        else:
-            await ctx.send('You can\'t create more than 50 or less than 1 channels at once')
+        if self.isAdmin(ctx.author):
+            arg = int(arg)
+            if arg <= 50 and arg >=1:
+                channels = getChannelNamesList(ctx.guild)
+                i = getMaxChannel(channels)
+                for _ in range(arg):
+                    await ctx.guild.create_text_channel(self._channelPrefix + str(i))
+                    i += 1
+                await ctx.send(f'{arg} channel(s) have been created.')
+            else:
+                await ctx.send('You can\'t create more than 50 or less than 1 channels at once')
 
     @commands.command()
     async def deleteChannel(self, ctx: commands.Context, arg=1):
-        arg = int(arg)
-        if arg <= 50 and arg >=1:
-            channels = getChannelNamesList(ctx.guild)
-            i = getMaxChannel(channels) - 1
-            for _ in range(arg):
-                for channel in ctx.guild.text_channels:
-                    if self._channelPrefix + str(i) == channel.name:
-                        await channel.delete()
-                        i -= 1
-                        break  
-            await ctx.send(f'{arg} channel(s) have been deleted.')
-        else:
-            await ctx.send('You can\'t delete more than 50 or less than 1 channels at once')
+        if self.isAdmin(ctx.author):
+            arg = int(arg)
+            if arg <= 50 and arg >=1:
+                channels = getChannelNamesList(ctx.guild)
+                i = getMaxChannel(channels) - 1
+                for _ in range(arg):
+                    for channel in ctx.guild.text_channels:
+                        if self._channelPrefix + str(i) == channel.name:
+                            await channel.delete()
+                            i -= 1
+                            break  
+                await ctx.send(f'{arg} channel(s) have been deleted.')
+            else:
+                await ctx.send('You can\'t delete more than 50 or less than 1 channels at once')
     ############# ENDOF CHANNELS RELATED #############
 
     ################ ROLES RELATED ################
     @commands.command()
     async def createRoles(self, ctx: commands.Context):
-        await ctx.guild.create_role(name="Bomb Party Admin")
-        await ctx.guild.create_role(name="BP Current Player")
+        if self.isAdmin(ctx.author):
+            await ctx.guild.create_role(name="Bomb Party Admin")
     ############# ENDOF ROLES RELATED #############
 
     ################ PARTY RELATED ################
@@ -149,64 +160,70 @@ class BombParty(commands.Cog):
         players = [] #e.g: [{"User": userInstance1, "Life": 2}, {"User": userInstance2, "Life": 2}, {"User": userInstance1, "Life": 1}]
         strPlayerList = 'Players:'
         unavailableWords = set()
-        async for message in ctx.channel.history(limit=100):
+        partyFindBool = False
+        async for message in ctx.channel.history(limit=200):
             if message.author == self.bot.user and message.content == "Party created, click on the reaction bellow to join!":
                 reac = message.reactions[0]
+                partyFindBool = True
                 break
-        async for user in reac.users():
-            if user != self.bot.user:
-                players.append({"User": user, "Life": 2})
-                strPlayerList += f' {user.mention} |'
-        await ctx.send(strPlayerList)
-        # for player in players:
-        #     player.add_roles()
-        end = False
-        Index = random.randint(0, len(players)-1)
-        if len(players) <= 1:
-                end = True
-        while(not end):
-            CurrentPlayer = players[Index]
-            #CurrentPlayer.add_role()
-            start = time.time()
-            letters = random.choice(self._dictionnaries[lang]["letters"])
-            await ctx.send(f'{CurrentPlayer["User"].mention}, type a word that contains: **{letters}**')
-            if timeLeft < minTiming:
-                timeLeft = minTiming
-            explode = True
-            while(timeLeft - (time.time() - start) > 0):
-                async for message in channel.history(limit=1):
-                    lastMsg = message
-                lastMsgContent = lastMsg.content.lower()
-                if lastMsg.author == CurrentPlayer["User"] and letters in lastMsgContent and lastMsgContent in self._dictionnaries[lang]["words"] and lastMsgContent not in unavailableWords:
-                    unavailableWords.add(lastMsgContent)
-                    explode = False
-                    break
-            timeLeft = timeLeft - (time.time() - start)
-            if explode:
-                timeLeft = baseTiming
-                CurrentPlayer["Life"] -= 1
-                if CurrentPlayer["Life"] <= 0:
-                    players.remove(CurrentPlayer)
-                    await ctx.send(f'💥BOOM💥, player {CurrentPlayer["User"].mention} haven\'t aswered as quickly enough!')
-                    if Index >= len(players) - 1:
-                        Index = 0
-                else:
-                    await ctx.send(f'💥BOOM💥, player {CurrentPlayer["User"].mention} has only {CurrentPlayer["Life"]} life remaining...')
-                    if Index >= len(players) - 1:
-                        Index = 0
+        if partyFindBool:
+            async for user in reac.users():
+                if user != self.bot.user:
+                    players.append({"User": user, "Life": 2})
+                    strPlayerList += f' {user.mention} |'
+            if len(players) > 1:
+                await ctx.send(strPlayerList)
+                # for player in players:
+                #     player.add_roles()
+                end = False
+                Index = random.randint(0, len(players)-1)
+                if len(players) <= 1:
+                        end = True
+                while(not end):
+                    CurrentPlayer = players[Index]
+                    #CurrentPlayer.add_role()
+                    start = time.time()
+                    letters = random.choice(self._dictionnaries[lang]["letters"])
+                    await ctx.send(f'{CurrentPlayer["User"].mention}, type a word that contains: **{letters}**')
+                    if timeLeft < minTiming:
+                        timeLeft = minTiming
+                    explode = True
+                    while(timeLeft - (time.time() - start) > 0):
+                        async for message in channel.history(limit=1):
+                            lastMsg = message
+                        lastMsgContent = lastMsg.content.lower()
+                        if lastMsg.author == CurrentPlayer["User"] and letters in lastMsgContent and lastMsgContent in self._dictionnaries[lang]["words"] and lastMsgContent not in unavailableWords:
+                            unavailableWords.add(lastMsgContent)
+                            explode = False
+                            break
+                    timeLeft = timeLeft - (time.time() - start)
+                    if explode:
+                        timeLeft = baseTiming
+                        CurrentPlayer["Life"] -= 1
+                        if CurrentPlayer["Life"] <= 0:
+                            players.remove(CurrentPlayer)
+                            await ctx.send(f'💥BOOM💥, player {CurrentPlayer["User"].mention} haven\'t aswered as quickly enough!')
+                            if Index >= len(players) - 1:
+                                Index = 0
+                        else:
+                            await ctx.send(f'💥BOOM💥, player {CurrentPlayer["User"].mention} has only {CurrentPlayer["Life"]} life remaining...')
+                            if Index >= len(players) - 1:
+                                Index = 0
+                            else:
+                                Index += 1
                     else:
-                        Index += 1
+                        if Index >= len(players) - 1:
+                            Index = 0
+                        else:
+                            Index += 1
+                    if len(players) == 1:
+                        end = True
+                await ctx.send(f'{players[0]["User"].mention} has won! 🏆')
             else:
-                if Index >= len(players) - 1:
-                    Index = 0
-                else:
-                    Index += 1
-            if len(players) == 1:
-                end = True
-        if len(players) == 0:
-            await ctx.send(f'This party has been canceled as not enough players joined...')
+                await ctx.send(f'This party has been canceled as not enough players joined...')
         else:
-            await ctx.send(f'{players[0]["User"].mention} has won! 🏆')
+            await ctx.send(f'No party have been created.')
+            
     ############# ENDOF PARTY RELATED #############
 
 def setup(bot):
